# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.DiscordGuild:
    properties:
      bot_addable:
        type: boolean
      bot_arrived:
        type: boolean
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      owner:
        type: boolean
      permissions:
        example: "0"
        type: string
    type: object
  entities.ListMyGuildsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entities.DiscordGuild'
        type: array
    type: object
...
  model.DiscordGuildStat:
    properties:
      created_at:
        type: string
      guild_id:
        type: string
      id:
        $ref: '#/definitions/uuid.NullUUID'
      nr_of_animated_emojis:
        type: integer
      nr_of_announcement_channels:
        type: integer
      nr_of_bots:
        type: integer
      nr_of_categories:
        type: integer
      nr_of_channels:
        type: integer
      nr_of_custom_stickers:
        type: integer
      nr_of_emojis:
        type: integer
      nr_of_members:
        type: integer
      nr_of_roles:
        type: integer
      nr_of_server_stickers:
        type: integer
      nr_of_stage_channels:
        type: integer
      nr_of_static_emojis:
        type: integer
      nr_of_stickers:
        type: integer
      nr_of_text_channels:
        type: integer
      nr_of_users:
        type: integer
      nr_of_voice_channels:
        type: integer
    type: object
...
  request.CreateGuildRequest:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  request.UpdateGuildRequest:
    properties:
      global_xp:
        type: string
      log_channel:
        type: string
    type: object
...
  response.GetGuildResponse:
    properties:
      alias:
        type: string
      bot_scopes:
        items:
          type: string
        type: array
      global_xp:
        type: boolean
      id:
        type: string
      log_channel:
        type: string
      log_channel_id:
        type: string
      name:
        type: string
    type: object
...
  response.GetGuildsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.GetGuildResponse'
        type: array
    type: object
...
  response.ResponseMessage:
    properties:
      message:
        type: string
    type: object
...
  uuid.NullUUID:
    properties:
      uuid:
        type: string
      valid:
        description: Valid is true if UUID is not NULL
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a swagger for mochi api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger API
  version: "1.0"
paths:
...
  /guilds:
    get:
      consumes:
      - application/json
      description: Get guilds
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetGuildsResponse'
      summary: Get guilds
      tags:
      - Guild
    post:
      consumes:
      - application/json
      description: Create guild
      parameters:
      - description: Create Guild request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.CreateGuildRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/request.CreateGuildRequest'
      summary: Create guild
      tags:
      - Guild
  /guilds/{guild_id}:
    get:
      consumes:
      - application/json
      description: Get guild
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetGuildResponse'
      summary: Get guild
      tags:
      - Guild
    put:
      consumes:
      - application/json
      description: Update guild
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      - description: Update guild request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateGuildRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseMessage'
      summary: Update guild
      tags:
      - Guild
  /guilds/{guild_id}/channels:
    post:
      consumes:
      - application/json
      description: Create guild channel
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      - description: Guild ID
        in: query
        name: count_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Create guild channel
      tags:
      - Guild
...
  /guilds/{guild_id}/stats:
    get:
      consumes:
      - application/json
      description: Get guild stats
      parameters:
      - description: Guild ID
        in: path
        name: guild_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DiscordGuildStat'
      summary: Get guild stats
      tags:
      - Guild
...
  /guilds/user-managed:
    get:
      consumes:
      - application/json
      description: Get my guild list
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ListMyGuildsResponse'
      summary: Get my guilds list
      tags:
      - Guild
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: access_token
    type: apiKey
swagger: "2.0"